+++ include/ironbee/engine_types.h
@@ -74,7 +74,6 @@ extern const ib_default_string_t ib_default_string;
 
 /* Public type declarations */
 typedef struct ib_conn_t ib_conn_t;
-typedef struct ib_txdata_t ib_txdata_t;
 typedef struct ib_tx_t ib_tx_t;
 typedef struct ib_tfn_t ib_tfn_t;
 typedef struct ib_logevent_t ib_logevent_t;
@@ -166,12 +165,6 @@ typedef enum ib_block_method_t {
     IB_BLOCK_METHOD_STATUS, /**< Block using an HTTP status code response. */
     IB_BLOCK_METHOD_CLOSE,  /**< Block by closing the transport (TCP) layer. */
 } ib_block_method_t;
-
-/** Transaction Data Structure */
-struct ib_txdata_t {
-    size_t              dlen;            /**< Data buffer length */
-    uint8_t            *data;            /**< Data buffer */
-};
 
 /** Connection Structure */
 struct ib_conn_t {