+++ include/ironbee/field.h
@@ -772,12 +772,12 @@ ib_status_t DLL_PUBLIC ib_field_copy(
  * @returns Status code
  */
 ib_status_t DLL_PUBLIC ib_field_create_bytestr_alias(
-    ib_field_t **pf,
-    ib_mpool_t  *mp,
-    const char  *name,
-    size_t       nlen,
-    uint8_t     *val,
-    size_t       vlen
+    ib_field_t    **pf,
+    ib_mpool_t     *mp,
+    const char     *name,
+    size_t          nlen,
+    const uint8_t  *val,
+    size_t          vlen
 );
 
 /**