+++ include/ironbeepp/all.hpp
@@ -57,7 +57,6 @@
 #include <ironbeepp/test_fixture.hpp>
 #include <ironbeepp/throw.hpp>
 #include <ironbeepp/transaction.hpp>
-#include <ironbeepp/transaction_data.hpp>
 #include <ironbeepp/transformation.hpp>
 #include <ironbeepp/var.hpp>